requirements.txt
Flask 3.0.2
Flask-Bcrypt 1.0.1
Flask-Login 0.6.3
Flask-Admin 1.6.1
Flask_Mail 0.9.1
Flask_SQLalchemy 3.1.1
